多语言init

+++ src/server/api/controller/upload.js
@@ -45,6 +45,10 @@ module.exports = {
   },
   delImg(req, res) {
     global.logger.info('resouce/delImg.json');
+    if (req.session.user.access > 1) {
+      res.json({ errno: 1, mes: '没有权限' })
+      return;
+    }
     var _name = req.query;
     fs.unlink(`${UPLOAD_IMG}/${_name.name}`, (err, f) => {
       if (err) {
@@ -55,6 +59,10 @@ module.exports = {
   },
   delDir(req, res) {
     global.logger.info('resouce/delDir.json');
+    if (req.session.user.access > 1) {
+      res.json({ errno: 1, mes: '没有权限' })
+      return;
+    }
     var _name = req.query;
     rimraf(`${UPLOAD_IMG}/${_name.name}`, (err, f) => {
       if (err) {
@@ -65,6 +73,10 @@ module.exports = {
   },
   rename(req, res) {
     global.logger.info('resouce/rename.json');
+    if (req.session.user.access > 1) {
+      res.json({ errno: 1, mes: '没有权限' })
+      return;
+    }
     var {origin, newname} = req.body;
     fs.rename(`${UPLOAD_IMG}/${origin}`, `${UPLOAD_IMG}/${newname}`, err => {
       if (err) {
@@ -74,6 +86,10 @@ module.exports = {
     })
   },
   mkdir(req, res) {
+    if (req.session.user.access > 1) {
+      res.json({ errno: 1, mes: '没有权限' })
+      return;
+    }
     var {dirName} = req.body;
     fs.mkdir(`${UPLOAD_IMG}/${dirName}`, '0777', function (err) {
       if (err) {
